--- ivy_tests/test_ivy/test_frontends/test_numpy/test_statistics/test_percentile.py
@@ -1,60 +1,0 @@
-# global
-from hypothesis import strategies as st, assume
-
-# local
-import ivy_tests.test_ivy.helpers as helpers
-from ivy_tests.test_ivy.test_functional.test_core.test_statistical import (
-    statistical_dtype_values,
-)
-import ivy_tests.test_ivy.test_frontends.test_numpy.helpers as np_frontend_helpers
-from ivy_tests.test_ivy.helpers import handle_frontend_test
-
-
-@handle_frontend_test(
-    fn_tree="numpy.percentile",
-    dtype_and_x=statistical_dtype_values(function="percentile"),
-    dtype=helpers.get_dtypes("float", full=False, none=True),
-    where=np_frontend_helpers.where(),
-    keep_dims=st.booleans(),
-)
-def test_numpy_percentile(
-    dtype_and_x,
-    dtype,
-    where,
-    as_variable,
-    with_out,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-    on_device,
-    keep_dims,
-):
-    input_dtype, x, axis = dtype_and_x
-    if isinstance(axis, tuple):
-        axis = axis[0]
-
-    where, as_variable, native_array = np_frontend_helpers.handle_where_and_array_bools(
-        where=where,
-        input_dtype=input_dtype,
-        as_variable=as_variable,
-        native_array=native_array,
-    )
-
-    np_frontend_helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=with_out,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-        axis=axis,
-        dtype=dtype[0],
-        out=None,
-        keepdims=keep_dims,
-        where=where,
-        test_values=False,
-    )
